requirements.txt
fastapi
uvicorn
Pillow
numpy
matplotlib
starlette
opencv-python
python-multipart
